+++ perl/FSM/Extension/Loader.pm
@@ -26,8 +26,7 @@ sub module_names_from_config_files ($self, $config_paths) {
 
 sub module_names_from_config_file ($self, $config_path) {
     return _with_extension_config_context($config_path, sub {
-        confess "FSM::Extension::Loader requires non-empty extension config paths"
-            unless defined($config_path) && $config_path ne '';
+        _validate_extension_config_path($config_path);
         confess "Extension config file not found: $config_path"
             unless -f $config_path;
 
@@ -66,10 +65,7 @@ sub load_modules ($self, $module_names) {
     my @extensions;
     for my $module_name (@$module_names) {
         my $extension = _with_extension_module_context($module_name, sub {
-            confess "FSM::Extension::Loader requires non-empty extension module names"
-                unless defined($module_name) && $module_name ne '';
-            confess "FSM::Extension::Loader rejects invalid extension module name '$module_name'"
-                unless _is_valid_extension_module_name($module_name);
+            _validate_extension_module_name($module_name);
 
             my $loaded = eval "require $module_name; 1";
             confess "Unable to load extension module '$module_name': $@"
@@ -94,6 +90,17 @@ sub load_modules ($self, $module_names) {
     return \@extensions;
 }
 
+sub _validate_extension_config_path ($config_path) {
+    die "FSM::Extension::Loader expects extension config paths to be scalar non-empty filesystem paths"
+        unless defined($config_path) && !ref($config_path) && $config_path =~ /\S/;
+}
+
+sub _validate_extension_module_name ($module_name) {
+    die "FSM::Extension::Loader expects extension module names to be scalar Module::Name values"
+        unless defined($module_name) && !ref($module_name) && $module_name =~ /\S/;
+    die "FSM::Extension::Loader rejects invalid extension module name '$module_name'"
+        unless _is_valid_extension_module_name($module_name);
+}
 sub _is_valid_extension_module_name ($module_name) {
     return 0 unless defined($module_name) && !ref($module_name);
     return $module_name =~ /\A[A-Za-z_]\w*(?:::[A-Za-z_]\w*)*\z/ ? 1 : 0;
@@ -109,7 +116,7 @@ sub _with_extension_config_context ($config_path, $code_ref) {
     die $error if ref($error);
     die $error if $error =~ /(?:^|\n)Extension config file:\s+'/s;
 
-    my $label = defined($config_path) && $config_path ne '' ? $config_path : '<missing>';
+    my $label = _extension_config_context_label($config_path);
     die "Extension config file: '$label'\n$error";
 }
 
@@ -123,8 +130,20 @@ sub _with_extension_module_context ($module_name, $code_ref) {
     die $error if ref($error);
     die $error if $error =~ /(?:^|\n)Extension module:\s+'/s;
 
-    my $label = defined($module_name) && $module_name ne '' ? $module_name : '<missing>';
+    my $label = _extension_module_context_label($module_name);
     die "Extension module: '$label'\n$error";
+}
+
+sub _extension_config_context_label ($config_path) {
+    return '<missing>' unless defined $config_path;
+    return '<non-scalar>' if ref($config_path);
+    return $config_path =~ /\S/ ? $config_path : '<blank>';
+}
+
+sub _extension_module_context_label ($module_name) {
+    return '<missing>' unless defined $module_name;
+    return '<non-scalar>' if ref($module_name);
+    return $module_name =~ /\S/ ? $module_name : '<blank>';
 }
 
 1;